Added zlib support for Win32 build. Since the MiniZip code (in main/zip/*) has _major_ C++ incompatibilities, I had to switch project file from (C++ compiler) back to Default. This also required refactoring all the code with the nice C99 features which MSVC stupidly doesn't support

+++ src/r4300/cop1_s.c
@@ -68,7 +68,7 @@ void SQRT_S(void)
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *reg_cop1_simple[cffd] = sqrt(*reg_cop1_simple[cffs]);
+   *reg_cop1_simple[cffd] = (float) sqrt(*reg_cop1_simple[cffs]);
    PC++;
 }
 
@@ -76,7 +76,7 @@ void ABS_S(void)
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *reg_cop1_simple[cffd] = fabs(*reg_cop1_simple[cffs]);
+   *reg_cop1_simple[cffd] = (float) fabs(*reg_cop1_simple[cffs]);
    PC++;
 }
